refactor: 옵션 관련 dto 별도 클래스로 분리
