+++ backend/src/main/java/me/eyeseeu/kiosk/option/dto/response/OptionCreateResponse.java
@@ -1,0 +1,10 @@
+package me.eyeseeu.kiosk.option.dto.response;
+
+public record OptionCreateResponse(
+    Long id,
+    String name,
+    int price,
+    String picture
+) {
+
+}
