+++ backend/src/main/java/me/eyeseeu/kiosk/option/dto/response/OptionGetResponse.java
@@ -1,0 +1,9 @@
+package me.eyeseeu.kiosk.option.dto.response;
+
+public record OptionGetResponse(
+    Long id,
+    String name,
+    int price,
+    String picture) {
+
+}
