+++ backend/src/main/java/me/eyeseeu/kiosk/option/dto/response/OptionGroupCreateResponse.java
@@ -10,13 +10,4 @@ public record OptionGroupCreateResponse(
     List<OptionCreateResponse> options
 ) {
 
-    public record OptionCreateResponse(
-        Long id,
-        String name,
-        int price,
-        String picture
-    ) {
-
-    }
-
 }
