+++ backend/src/main/java/me/eyeseeu/kiosk/option/dto/response/OptionGroupGetResponse.java
@@ -10,13 +10,4 @@ public record OptionGroupGetResponse(
     List<OptionGetResponse> options
 ) {
 
-    public record OptionGetResponse(
-        Long id,
-        String name,
-        int price,
-        String picture
-    ) {
-
-    }
-
 }
