+++ backend/src/main/java/me/eyeseeu/kiosk/option/service/OptionGroupService.java
@@ -9,9 +9,9 @@ import me.eyeseeu.kiosk.member.entity.Member;
 import me.eyeseeu.kiosk.member.service.MemberService;
 import me.eyeseeu.kiosk.option.dto.request.OptionGroupCreateRequest;
 import me.eyeseeu.kiosk.option.dto.response.OptionGroupCreateResponse;
-import me.eyeseeu.kiosk.option.dto.response.OptionGroupCreateResponse.OptionCreateResponse;
+import me.eyeseeu.kiosk.option.dto.response.OptionCreateResponse;
 import me.eyeseeu.kiosk.option.dto.response.OptionGroupGetResponse;
-import me.eyeseeu.kiosk.option.dto.response.OptionGroupGetResponse.OptionGetResponse;
+import me.eyeseeu.kiosk.option.dto.response.OptionGetResponse;
 import me.eyeseeu.kiosk.option.entity.Option;
 import me.eyeseeu.kiosk.option.entity.OptionGroup;
 import me.eyeseeu.kiosk.option.repository.OptionGroupRepository;
